Fix NAG compiler issues in GMAO_gfio and GMAO_hermes
- GMAO_gfio/CMakeLists.txt: add -wmismatch for legacy NetCDF-2 C API
  routines (NCVDEF, NCAPT, NCVPT, NCAGT, NCVGT, NCVGT1) called with
  varying Fortran types intentionally
- GMAO_hermes/CMakeLists.txt: remove duplicate ${MISMATCH} from
  CMAKE_Fortran_FLAGS_RELEASE (already in common_Fortran_flags); add
  -wmismatch for GFIO_PUTREALATT/GFIO_GETREALATT type-punning
- GMAO_hermes/m_nc_JEDIinc.f90: replace byte-count kind literal real(4)
  with portable real(selected_real_kind(6)) via parameter r4
- GMAO_hermes/m_nc_akbk.f90: replace double precision with real(real64)
  from iso_fortran_env so NF90_PUT_VAR generic resolves correctly
  when compiled with -r8

+++ GMAO_hermes/m_nc_akbk.f90
@@ -1,5 +1,6 @@
 module m_nc_akbk
 use netcdf
+use iso_fortran_env, only: real64
 implicit none
 private
 public :: write_nc_akbk
@@ -12,14 +13,14 @@ contains
 subroutine write_nc_akbk_ (fname,ak, bk)
 
   character(len=*), intent(in) :: fname
-  double precision, intent(in) :: ak(:), bk(:)
+  real(real64), intent(in) :: ak(:), bk(:)
 
   ! NetCDF variables
   integer :: ncid, dimid_edge
   integer :: varid_edge, varid_ak, varid_bk
   integer :: rc
   integer :: i, km
-  double precision, allocatable :: edge(:)
+  real(real64), allocatable :: edge(:)
   
   km = size(ak)
   allocate(edge(km))
